feat: Add AddNewUserButton component for user management
