+++ apps/studio/src/features/users/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./UserTableTabs"
+export * from "./AddNewUserButton"
